refactor: 설명 추가
